+++ src/Query/Builder.php
@@ -792,10 +792,38 @@ class Builder extends BaseBuilder
         return $this->performUpdate($query, $options);
     }
 
+    public function incrementEach(array $columns, array $extra = [], array $options = [])
+    {
+        $stage['$addFields'] = $extra;
+
+        // Not using $inc for each column, because it would fail if one column is null.
+        foreach ($columns as $column => $amount) {
+            $stage['$addFields'][$column] = [
+                '$add' => [$amount, ['$ifNull' => ['$' . $column, 0]]],
+            ];
+        }
+
+        $options = $this->inheritConnectionOptions($options);
+
+        return $this->performUpdate([$stage], $options);
+    }
+
     /** @inheritdoc */
     public function decrement($column, $amount = 1, array $extra = [], array $options = [])
     {
         return $this->increment($column, -1 * $amount, $extra, $options);
+    }
+
+    /** @inheritdoc */
+    public function decrementEach(array $columns, array $extra = [], array $options = [])
+    {
+        $decrement = [];
+
+        foreach ($columns as $column => $amount) {
+            $decrement[$column] = -1 * $amount;
+        }
+
+        return $this->incrementEach($decrement, $extra, $options);
     }
 
     /** @inheritdoc */
@@ -1243,7 +1271,8 @@ class Builder extends BaseBuilder
                 // All backslashes are converted to \\, which are needed in matching regexes.
                 preg_quote($value),
             );
-            $value = new Regex('^' . $regex . '$', 'i');
+            $flags = $where['caseSensitive'] ?? false ? '' : 'i';
+            $value = new Regex('^' . $regex . '$', $flags);
 
             // For inverse like operations, we can just use the $not operator with the Regex
             $operator = $operator === 'like' ? '=' : 'not';
@@ -1299,6 +1328,13 @@ class Builder extends BaseBuilder
     protected function compileWhereNotIn(array $where): array
     {
         return [$where['column'] => ['$nin' => array_values($where['values'])]];
+    }
+
+    protected function compileWhereLike(array $where): array
+    {
+        $where['operator'] = $where['not'] ? 'not like' : 'like';
+
+        return $this->compileWhereBasic($where);
     }
 
     protected function compileWhereNull(array $where): array